+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Changelog
 New Features
 ============
 - Add no_labels option to waves visualize (:issue:`583`, :merge:`717`). By `Prabhu Khalsa`_.
+- New print-tree feature in waves visualize (:issue:`582`, :merge:`718`). By `Prabhu Khalsa`_.
 
 Bug fixes
 =========